+++ src/plone/api/tests/test_content.py
@@ -754,6 +754,20 @@ class TestPloneApiContent(unittest.TestCase):
             context=self.portal.about, depth=0, portal_type='Document')
         self.assertEqual(len(documents), 0)
 
+        # Limit search depth with explicit path
+        documents = api.content.find(
+            path='/'.join(self.portal.about.getPhysicalPath()),
+            depth=1, portal_type='Document')
+        self.assertEqual(len(documents), 2)
+        documents = api.content.find(
+            path='/'.join(self.portal.about.getPhysicalPath()),
+            depth=0, portal_type='Document')
+        self.assertEqual(len(documents), 0)
+        documents = api.content.find(
+            path='/'.join(self.portal.events.getPhysicalPath()),
+            depth=1, portal_type='Document')
+        self.assertEqual(len(documents), 0)
+
     def test_find_interface(self):
         # Find documents by interface or it's identifier
         identifier = IContentish.__identifier__
@@ -779,6 +793,24 @@ class TestPloneApiContent(unittest.TestCase):
         query = {
             'portal_type': 'Document',
             'path': {'query': path, 'depth': 0}
+            }
+        documents = api.content.find(**query)
+        self.assertEqual(len(documents), 0)
+
+        # This is a bit awkward, but it is nice if this does not crash.
+        query = {
+            'depth': 2,
+            'portal_type': 'Document',
+            'path': {'query': path}
+            }
+        documents = api.content.find(**query)
+        self.assertEqual(len(documents), 2)
+
+        path = '/'.join(self.portal.events.getPhysicalPath())
+        query = {
+            'depth': 2,
+            'portal_type': 'Document',
+            'path': {'query': path}
             }
         documents = api.content.find(**query)
         self.assertEqual(len(documents), 0)